削除確認アラート追加

+++ src/app/childName/[childName]/page.tsx
@@ -239,9 +239,15 @@ const ChildReservationPage = () => {
 
   const handleDelete = async (postIndex: number) => {
     const postToDelete = filteredPosts[postIndex];
+
     if (postToDelete && postToDelete.id) {
-      await deleteDoc(doc(db, "posts", postToDelete.id));
-      setShouldFetch(true);
+      // 削除確認ダイアログを表示
+      const confirmed = window.confirm("本当に削除しますか？");
+
+      if (confirmed) {
+        await deleteDoc(doc(db, "posts", postToDelete.id));
+        setShouldFetch(true); // データを再取得するためのフラグを立てる
+      }
     }
   };
